Se agregaron los filtros Roberts y Frei-Chen

+++ Src/matrix.cpp
@@ -4,10 +4,10 @@ Matrix::Matrix(int rows, int columns)
 {
     this->rows = rows;
     this->columns = columns;
-    this->values = new int*[rows];
+    this->values = new float*[rows];
     for (int i=0; i < rows; i++)
     {
-        this->values[i] = new int[columns];
+        this->values[i] = new float[columns];
         for (int j=0; j < columns; j++)
             this->values[i][j] = 0;
     }
@@ -18,7 +18,7 @@ int Matrix::at(int r, int c)
    return this->values[r][c];
 }
 
-void Matrix::set(int r, int c, int value)
+void Matrix::set(int r, int c, float value)
 {
     this->values[r][c] = value;
 }
